bot_async_factory/telegram/requirements.txt
python-telegram-bot 
redis 
asyncio 
python-dotenv 
loguru